requirements.txt
Django==2.0.4
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.1
pytz==2018.4
